ATmega32: merge work from simpler-api to split up chip config, standardize exports around
