+++ src/peripherals/gpio_atmega32.ts
@@ -1,27 +1,27 @@
 import { AVRPortConfig } from './gpio';
 
-export const atmega32PortAConfig: AVRPortConfig = {
+export const portAConfig: AVRPortConfig = {
   PORT: 0x3b,
   DDR: 0x3a,
   PIN: 0x39,
   externalInterrupts: [],
 };
 
-export const atmega32PortBConfig: AVRPortConfig = {
+export const portBConfig: AVRPortConfig = {
   PORT: 0x38,
   DDR: 0x37,
   PIN: 0x36,
   externalInterrupts: [],
 };
 
-export const atmega32PortCConfig: AVRPortConfig = {
+export const portCConfig: AVRPortConfig = {
   PORT: 0x35,
   DDR: 0x34,
   PIN: 0x33,
   externalInterrupts: [],
 };
 
-export const atmega32PortDConfig: AVRPortConfig = {
+export const portDConfig: AVRPortConfig = {
   PORT: 0x32,
   DDR: 0x31,
   PIN: 0x30,
