+++ src/peripherals/timer.spec.ts
@@ -3,11 +3,7 @@ import { asmProgram, TestProgramRunner } from '../utils/test-utils';
 import { AVRIOPort, PinOverrideMode } from './gpio';
 import { portBConfig, portDConfig } from './gpio_atmega328p';
 import { AVRTimer } from './timer';
-import {
-  atmega328pTimer0Config,
-  atmega328pTimer1Config,
-  atmega328pTimer2Config,
-} from './timer_atmega328p';
+import { timer0Config, timer1Config, timer2Config } from './timer_atmega328p';
 
 // CPU registers
 const R1 = 1;
@@ -84,7 +80,7 @@ const nopOpCode = '0000';
 describe('timer', () => {
   it('should update timer every tick when prescaler is 1', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCCR0B, CS00); // Set prescaler to 1
     cpu.cycles = 1;
     cpu.tick();
@@ -96,7 +92,7 @@ describe('timer', () => {
 
   it('should update timer every 64 ticks when prescaler is 3', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCCR0B, CS01 | CS00); // Set prescaler to 64
     cpu.cycles = 1;
     cpu.tick();
@@ -108,7 +104,7 @@ describe('timer', () => {
 
   it('should not update timer if it has been disabled', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCCR0B, 0); // No prescaler (timer disabled)
     cpu.cycles = 1;
     cpu.tick();
@@ -120,7 +116,7 @@ describe('timer', () => {
 
   it('should set the TOV flag when timer wraps above TOP value', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCNT0, 0xff);
     cpu.writeData(TCCR0B, CS00); // Set prescaler to 1
 
@@ -137,7 +133,7 @@ describe('timer', () => {
 
   it('should set the TOV if timer overflows past TOP without reaching TOP', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCNT0, 0xfe);
     cpu.writeData(TCCR0B, CS00); // Set prescaler to 1
     cpu.cycles = 1;
@@ -151,7 +147,7 @@ describe('timer', () => {
 
   it('should clear the TOV flag when writing 1 to the TOV bit, and not trigger the interrupt', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCNT0, 0xff);
     cpu.writeData(TCCR0B, CS00); // Set prescaler to 1
     cpu.cycles = 1;
@@ -165,7 +161,7 @@ describe('timer', () => {
 
   it('should set TOV if timer overflows in FAST PWM mode', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCNT0, 0xff);
     cpu.writeData(TCCR0B, CS00); // Set prescaler to 1
     cpu.cycles = 1;
@@ -181,7 +177,7 @@ describe('timer', () => {
 
   it('should generate an overflow interrupt if timer overflows and interrupts enabled', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCNT0, 0xff);
     cpu.writeData(TCCR0B, CS00); // Set prescaler to 1
     cpu.cycles = 1;
@@ -200,7 +196,7 @@ describe('timer', () => {
   it('should support overriding TIFR/TOV and TIMSK/TOIE bits (issue #64)', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
     new AVRTimer(cpu, {
-      ...atmega328pTimer0Config,
+      ...timer0Config,
 
       // The following values correspond ATtiny85 config:
       TOV: 2,
@@ -227,7 +223,7 @@ describe('timer', () => {
 
   it('should not generate an overflow interrupt when global interrupts disabled', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCNT0, 0xff);
     cpu.writeData(TCCR0B, CS00); // Set prescaler to 1
     cpu.cycles = 1;
@@ -243,7 +239,7 @@ describe('timer', () => {
 
   it('should not generate an overflow interrupt when TOIE0 is clear', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCNT0, 0xff);
     cpu.writeData(TCCR0B, CS00); // Set prescaler to 1
     cpu.cycles = 1;
@@ -259,7 +255,7 @@ describe('timer', () => {
 
   it('should set OCF0A/B flags when OCRA/B == 0 and the timer equals to OCRA (issue #74)', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCNT0, 0xff);
     cpu.writeData(OCR0A, 0x0);
     cpu.writeData(OCR0B, 0x0);
@@ -277,7 +273,7 @@ describe('timer', () => {
 
   it('should set the OCF1A flag when OCR1A == 120 and the timer overflowed past 120 in WGM mode 15 (issue #94)', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer1Config);
+    new AVRTimer(cpu, timer1Config);
     cpu.writeData(TCNT1, 118);
     cpu.writeData(OCR1A, 120);
     cpu.writeData(OCR1B, 4); // To avoid getting the OCF1B flag set
@@ -295,7 +291,7 @@ describe('timer', () => {
 
   it('should set OCF0A flag when timer equals OCRA', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCNT0, 0x10);
     cpu.writeData(OCR0A, 0x11);
     cpu.writeData(TCCR0A, 0x0); // WGM: Normal
@@ -311,7 +307,7 @@ describe('timer', () => {
 
   it('should reset the counter in CTC mode if it equals to OCRA', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCNT0, 0x10);
     cpu.writeData(OCR0A, 0x11);
     cpu.writeData(TCCR0A, WGM01); // WGM: CTC
@@ -328,7 +324,7 @@ describe('timer', () => {
 
   it('should not set the TOV bit when TOP < MAX in CTC mode (issue #75)', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCNT0, 0x1e);
     cpu.writeData(OCR0A, 0x1f);
     cpu.writeData(TCCR0A, WGM01); // WGM: CTC
@@ -346,7 +342,7 @@ describe('timer', () => {
 
   it('should set the TOV bit when TOP == MAX in CTC mode (issue #75)', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCNT0, 0xfe);
     cpu.writeData(OCR0A, 0xff);
     cpu.writeData(TCCR0A, WGM01); // WGM: CTC
@@ -367,7 +363,7 @@ describe('timer', () => {
 
   it('should not set the TOV bit twice on overflow (issue #80)', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCNT0, 0xfe);
     cpu.writeData(OCR0A, 0xff);
     cpu.writeData(TCCR0A, WGM01); // WGM: CTC
@@ -389,7 +385,7 @@ describe('timer', () => {
 
   it('should set OCF0B flag when timer equals OCRB', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCNT0, 0x10);
     cpu.writeData(OCR0B, 0x11);
     cpu.writeData(TCCR0A, 0x0); // WGM: (Normal)
@@ -405,7 +401,7 @@ describe('timer', () => {
 
   it('should generate Timer Compare A interrupt when TCNT0 == TCNTA', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCNT0, 0x20);
     cpu.writeData(OCR0A, 0x21);
     cpu.writeData(TCCR0B, CS00); // Set prescaler to 1
@@ -424,7 +420,7 @@ describe('timer', () => {
 
   it('should not generate Timer Compare A interrupt when OCIEA is disabled', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCNT0, 0x20);
     cpu.writeData(OCR0A, 0x21);
     cpu.writeData(TCCR0B, CS00); // Set prescaler to 1
@@ -442,7 +438,7 @@ describe('timer', () => {
 
   it('should generate Timer Compare B interrupt when TCNT0 == TCNTB', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCNT0, 0x20);
     cpu.writeData(OCR0B, 0x21);
     cpu.writeData(TCCR0B, CS00); // Set prescaler to 1
@@ -471,7 +467,7 @@ describe('timer', () => {
       IN r17, 0x26    ; r17 <- TCNT
     `);
     const cpu = new CPU(program);
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     const runner = new TestProgramRunner(cpu);
     runner.runInstructions(instructionCount);
     expect(cpu.data[R17]).toEqual(0x31);
@@ -479,7 +475,7 @@ describe('timer', () => {
 
   it('timer2 should count every 256 ticks when prescaler is 6 (issue #5)', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer2Config);
+    new AVRTimer(cpu, timer2Config);
     cpu.writeData(TCCR2B, CS22 | CS21); // Set prescaler to 256
     cpu.cycles = 1;
     cpu.tick();
@@ -503,7 +499,7 @@ describe('timer', () => {
       LDS r1, 0x46      ; r1 <- TCNT0 (2 cycles)
     `);
     const cpu = new CPU(program);
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     const runner = new TestProgramRunner(cpu);
     runner.runInstructions(instructionCount);
     expect(cpu.data[R1]).toEqual(2);
@@ -521,7 +517,7 @@ describe('timer', () => {
       LDS r17, 0xb2   ; TCNT should equal 2 at this point
     `);
     const cpu = new CPU(program);
-    new AVRTimer(cpu, atmega328pTimer2Config);
+    new AVRTimer(cpu, timer2Config);
     const runner = new TestProgramRunner(cpu);
     runner.runInstructions(instructionCount);
     expect(cpu.readData(R17)).toEqual(2);
@@ -537,7 +533,7 @@ describe('timer', () => {
       LDS r17, 0xb2   ; TCNT2 should equal 2 at this point (not counting the NOP)
   `);
     const cpu = new CPU(program);
-    new AVRTimer(cpu, atmega328pTimer2Config);
+    new AVRTimer(cpu, timer2Config);
     const runner = new TestProgramRunner(cpu);
     runner.runInstructions(instructionCount);
     expect(cpu.readData(R17)).toEqual(2);
@@ -545,7 +541,7 @@ describe('timer', () => {
 
   it('should clear OC0B pin when writing 1 to FOC0B', () => {
     const cpu = new CPU(new Uint16Array(0x1000));
-    new AVRTimer(cpu, atmega328pTimer0Config);
+    new AVRTimer(cpu, timer0Config);
     cpu.writeData(TCCR0A, COM0B1);
 
     // Listen to Port B's internal callback
@@ -582,7 +578,7 @@ describe('timer', () => {
       `);
 
       const cpu = new CPU(program);
-      new AVRTimer(cpu, atmega328pTimer0Config);
+      new AVRTimer(cpu, timer0Config);
 
       // Listen to Port D's internal callback
       const portD = new AVRIOPort(cpu, portDConfig);
@@ -635,7 +631,7 @@ describe('timer', () => {
       `);
 
       const cpu = new CPU(program);
-      new AVRTimer(cpu, atmega328pTimer0Config);
+      new AVRTimer(cpu, timer0Config);
 
       // Listen to Port D's internal callback
       const portD = new AVRIOPort(cpu, portDConfig);
@@ -681,7 +677,7 @@ describe('timer', () => {
       `);
 
       const cpu = new CPU(program);
-      new AVRTimer(cpu, atmega328pTimer0Config);
+      new AVRTimer(cpu, timer0Config);
 
       // Listen to Port D's internal callback
       const portD = new AVRIOPort(cpu, portDConfig);
@@ -724,7 +720,7 @@ describe('timer', () => {
         IN r22, 0x26   ; TCNT0 will be 1 (end of test)
       `);
       const cpu = new CPU(program);
-      new AVRTimer(cpu, atmega328pTimer0Config);
+      new AVRTimer(cpu, timer0Config);
       const runner = new TestProgramRunner(cpu);
       runner.runInstructions(instructionCount);
 
@@ -755,7 +751,7 @@ describe('timer', () => {
       `);
 
       const cpu = new CPU(program);
-      new AVRTimer(cpu, atmega328pTimer0Config);
+      new AVRTimer(cpu, timer0Config);
 
       // Listen to Port D's internal callback
       const portD = new AVRIOPort(cpu, portDConfig);
@@ -802,7 +798,7 @@ describe('timer', () => {
       `);
 
       const cpu = new CPU(program);
-      new AVRTimer(cpu, atmega328pTimer0Config);
+      new AVRTimer(cpu, timer0Config);
 
       // Listen to Port D's internal callback
       const portD = new AVRIOPort(cpu, portDConfig);
@@ -835,7 +831,7 @@ describe('timer', () => {
       `);
 
       const cpu = new CPU(program);
-      new AVRTimer(cpu, atmega328pTimer0Config);
+      new AVRTimer(cpu, timer0Config);
 
       // Listen to Port D's internal callback
       const portD = new AVRIOPort(cpu, portDConfig);
@@ -863,7 +859,7 @@ describe('timer', () => {
       `);
 
       const cpu = new CPU(program);
-      new AVRTimer(cpu, atmega328pTimer0Config);
+      new AVRTimer(cpu, timer0Config);
 
       // Listen to Port D's internal callback
       const portD = new AVRIOPort(cpu, portDConfig);
@@ -906,7 +902,7 @@ describe('timer', () => {
       `);
 
       const cpu = new CPU(program);
-      new AVRTimer(cpu, atmega328pTimer0Config);
+      new AVRTimer(cpu, timer0Config);
 
       const runner = new TestProgramRunner(cpu);
       runner.runInstructions(instructionCount);
@@ -934,7 +930,7 @@ describe('timer', () => {
       `);
 
       const cpu = new CPU(program);
-      new AVRTimer(cpu, atmega328pTimer0Config);
+      new AVRTimer(cpu, timer0Config);
 
       const runner = new TestProgramRunner(cpu);
       runner.runInstructions(instructionCount);
@@ -958,7 +954,7 @@ describe('timer', () => {
       `);
 
       const cpu = new CPU(program);
-      const timer = new AVRTimer(cpu, atmega328pTimer0Config);
+      const timer = new AVRTimer(cpu, timer0Config);
 
       const runner = new TestProgramRunner(cpu);
       runner.runInstructions(instructionCount);
@@ -971,7 +967,7 @@ describe('timer', () => {
   describe('16 bit timers', () => {
     it('should increment 16-bit TCNT by 1', () => {
       const cpu = new CPU(new Uint16Array(0x1000));
-      new AVRTimer(cpu, atmega328pTimer1Config);
+      new AVRTimer(cpu, timer1Config);
       cpu.writeData(TCNT1H, 0x22); // TCNT1 <- 0x2233
       cpu.writeData(TCNT1, 0x33); // ...
       const timerLow = cpu.readData(TCNT1);
@@ -989,7 +985,7 @@ describe('timer', () => {
 
     it('should set OCF0A flag when timer equals OCRA (16 bit mode)', () => {
       const cpu = new CPU(new Uint16Array(0x1000));
-      new AVRTimer(cpu, atmega328pTimer1Config);
+      new AVRTimer(cpu, timer1Config);
       cpu.writeData(TCNT1H, 0x10); // TCNT1 <- 0x10ee
       cpu.writeData(TCNT1, 0xee); // ...
       cpu.writeData(OCR1AH, 0x10); // OCR1 <- 0x10ef
@@ -1011,7 +1007,7 @@ describe('timer', () => {
       const OCR1CH = 0x8d;
       const OCF1C = 1 << 3;
       new AVRTimer(cpu, {
-        ...atmega328pTimer1Config,
+        ...timer1Config,
         OCRC: OCR1C,
         OCFC: OCF1C,
       });
@@ -1032,7 +1028,7 @@ describe('timer', () => {
 
     it('should generate an overflow interrupt if timer overflows and interrupts enabled', () => {
       const cpu = new CPU(new Uint16Array(0x1000));
-      new AVRTimer(cpu, atmega328pTimer1Config);
+      new AVRTimer(cpu, timer1Config);
       cpu.writeData(TCCR1A, 0x3); // TCCR1A <- WGM10 | WGM11 (Fast PWM, 10-bit)
       cpu.writeData(TCCR1B, 0x9); // TCCR1B <- WGM12 | CS10
       cpu.writeData(TIMSK1, 0x1); // TIMSK1: TOIE1
@@ -1054,7 +1050,7 @@ describe('timer', () => {
 
     it('should reset the timer once it reaches ICR value in mode 12', () => {
       const cpu = new CPU(new Uint16Array(0x1000));
-      new AVRTimer(cpu, atmega328pTimer1Config);
+      new AVRTimer(cpu, timer1Config);
       cpu.writeData(TCNT1H, 0x50); // TCNT1 <- 0x500f
       cpu.writeData(TCNT1, 0x0f); // ...
       cpu.writeData(ICR1H, 0x50); // ICR1 <- 0x5010
@@ -1072,7 +1068,7 @@ describe('timer', () => {
 
     it('should not update the high byte of TCNT if written after the low byte (issue #37)', () => {
       const cpu = new CPU(new Uint16Array(0x1000));
-      new AVRTimer(cpu, atmega328pTimer1Config);
+      new AVRTimer(cpu, timer1Config);
       cpu.writeData(TCNT1, 0x22);
       cpu.writeData(TCNT1H, 0x55);
       cpu.cycles = 1;
@@ -1084,7 +1080,7 @@ describe('timer', () => {
 
     it('reading from TCNT1H before TCNT1L should return old value (issue #37)', () => {
       const cpu = new CPU(new Uint16Array(0x1000));
-      new AVRTimer(cpu, atmega328pTimer1Config);
+      new AVRTimer(cpu, timer1Config);
       cpu.writeData(TCNT1H, 0xff);
       cpu.writeData(TCNT1, 0xff);
       cpu.writeData(TCCR1B, WGM12 | CS10); // Set prescaler to 1, WGM: CTC
@@ -1120,7 +1116,7 @@ describe('timer', () => {
       `);
 
       const cpu = new CPU(program);
-      new AVRTimer(cpu, atmega328pTimer1Config);
+      new AVRTimer(cpu, timer1Config);
 
       // Listen to Port B's internal callback
       const portB = new AVRIOPort(cpu, portBConfig);
@@ -1161,7 +1157,7 @@ describe('timer', () => {
 
       const cpu = new CPU(program);
       new AVRTimer(cpu, {
-        ...atmega328pTimer1Config,
+        ...timer1Config,
         OCRC: OCR1C,
         OCFC: OCF1C,
         compPortC: portBConfig.PORT,
@@ -1188,7 +1184,7 @@ describe('timer', () => {
     it('should toggle OC1C on when writing 1 to FOC1C', () => {
       const cpu = new CPU(new Uint16Array(0x1000));
       new AVRTimer(cpu, {
-        ...atmega328pTimer1Config,
+        ...timer1Config,
         OCRC: OCR1C,
         OCFC: OCF1C,
         compPortC: portBConfig.PORT,
@@ -1208,7 +1204,7 @@ describe('timer', () => {
     it('should not toggle OC1C on when writing 1 to FOC1C in PWM mode', () => {
       const cpu = new CPU(new Uint16Array(0x1000));
       new AVRTimer(cpu, {
-        ...atmega328pTimer1Config,
+        ...timer1Config,
         OCRC: OCR1C,
         OCFC: OCF1C,
         compPortC: portBConfig.PORT,
@@ -1257,7 +1253,7 @@ describe('timer', () => {
       `);
 
       const cpu = new CPU(program);
-      new AVRTimer(cpu, atmega328pTimer1Config);
+      new AVRTimer(cpu, timer1Config);
 
       const runner = new TestProgramRunner(cpu);
       runner.runInstructions(instructionCount);
@@ -1293,7 +1289,7 @@ describe('timer', () => {
       `);
 
       const cpu = new CPU(program);
-      new AVRTimer(cpu, atmega328pTimer1Config);
+      new AVRTimer(cpu, timer1Config);
 
       const runner = new TestProgramRunner(cpu);
       runner.runInstructions(instructionCount);
@@ -1306,7 +1302,7 @@ describe('timer', () => {
 
     it('should mask the unused bits of OCR1A when using fixed top values', () => {
       const cpu = new CPU(new Uint16Array(0x1000));
-      new AVRTimer(cpu, atmega328pTimer1Config);
+      new AVRTimer(cpu, timer1Config);
       cpu.writeData(TCCR1A, WGM10 | WGM11); // WGM: FastPWM, top 0x3ff
       cpu.writeData(TCCR1B, WGM12);
       cpu.writeData(OCR1AH, 0xff);
@@ -1320,7 +1316,7 @@ describe('timer', () => {
     it('should count on the falling edge of T0 when CS=110', () => {
       const cpu = new CPU(new Uint16Array(0x1000));
       const port = new AVRIOPort(cpu, portDConfig);
-      new AVRTimer(cpu, atmega328pTimer0Config);
+      new AVRTimer(cpu, timer0Config);
       cpu.writeData(TCCR0B, CS02 | CS01); // Count on falling edge
       cpu.cycles = 1;
       cpu.tick();
@@ -1339,7 +1335,7 @@ describe('timer', () => {
     it('should count on the rising edge of T0 when CS=111', () => {
       const cpu = new CPU(new Uint16Array(0x1000));
       const port = new AVRIOPort(cpu, portDConfig);
-      new AVRTimer(cpu, atmega328pTimer0Config);
+      new AVRTimer(cpu, timer0Config);
       cpu.writeData(TCCR0B, CS02 | CS01 | CS00); // Count on rising edge
       cpu.cycles = 1;
       cpu.tick();
